fix cli tests conflicting with non-cli tests
